RahafFinalPosh

+++ MusouqSystem/src/main/java/com/example/musouqsystem/Service/ShippingCompanyService.java
@@ -2,6 +2,7 @@ package com.example.musouqsystem.Service;
 
 import com.example.musouqsystem.Api.ApiException;
 import com.example.musouqsystem.Model.ShippingCompany;
+import com.example.musouqsystem.Repository.OrdersRepository;
 import com.example.musouqsystem.Repository.ShippingCompanyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ShippingCompanyService {
     private final ShippingCompanyRepository shippingCompanyRepository;
+    private final OrdersRepository ordersRepository;
 
     public List<ShippingCompany> shopperGetAllShippingCompany() {
         return shippingCompanyRepository.findAll();
@@ -34,6 +36,8 @@ public class ShippingCompanyService {
     public void adminDeleteShippingCompany(Integer shippingCompany_id) {
         ShippingCompany shippingCompany1 = shippingCompanyRepository.findShippingCompanyById(shippingCompany_id);
         if (shippingCompany1 == null) throw new ApiException("The shippingCompany not found");
+        if(!shippingCompany1.getOrders().isEmpty())
+            throw new ApiException("you can't delete this company becuase there is Orders related with it");
         shippingCompanyRepository.delete(shippingCompany1);
     }
